Update README.md file for building this project.
This is a very simple example, which is only suitable for developers who want to lean gRPC briefly and quickly, not suitable for profound research.
You can get more in master branch, if you want to know more.
This will be the last commit in this branch, I'll stop updating later.